Updated tileset to have a side heads.
Compressed the tileset.png win ImageMagick so it uses only RGB colores.
Reorganized tileset.png to make big heads without changing the current sprites rendering logic.
Optimized code for turning the head around.
Added utility function to wrap a number (might be useful for looping... like the head turning).
Created variables to match the head with its side when creating robots. This can yet be optimized.
Define a rule to override colors in the head when it is looking backwards.